tests/dobbs_testing/requirements.txt
# Requirements for Smart Chunking Test Script
# Install with: pip install -r requirements.txt

PyYAML>=6.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0